Rename player_progress to player_data.
