+++ lib/widgets/game_screen/game_screen.dart
@@ -6,7 +6,7 @@ import 'package:elola/localizations.dart';
 import 'package:elola/configs/constants.dart' as constants;
 import 'package:elola/models/noun.dart';
 import 'package:elola/modules/noun_database/noun_database.dart';
-import 'package:elola/modules/player_progress/player_progress.dart';
+import 'package:elola/modules/player_data/player_data.dart';
 import 'package:elola/modules/text_to_speech/text_to_speech.dart';
 
 class GameScreen extends StatefulWidget {
